Small tidy up of content controller
