+++ server/controllers/content.js
@@ -1,8 +1,13 @@
 import Prismic from 'prismic-javascript';
 import {prismicApi} from '../services/prismic-api';
-import {getEditorial, getEditorialPreview, getEvent} from '../services/prismic-content';
+import {getEditorial, getEditorialList, getEditorialPreview, getEvent} from '../services/prismic-content';
 import {createPageConfig} from '../model/page-config';
 import {getEventbriteEventEmbed} from '../services/eventbrite';
+import {PromoFactory} from '../model/promo';
+import {getArticleStubs} from '../services/wordpress';
+import {getCuratedList} from '../services/prismic-curated-lists';
+import {isFlagEnabled} from '../util/flag-status';
+import {collectorsPromo} from '../data/series';
 
 export const renderEditorial = async(ctx, next) => {
   const format = ctx.request.query.format;
@@ -93,3 +98,36 @@ export const renderEventbriteEmbed = async(ctx, next) => {
 
   return next();
 };
+
+export async function renderExplore(ctx, next) {
+  // TODO: Remove WP content
+  const [flags] = ctx.intervalCache.get('flags');
+  const prismicArticlesOnExploreFlag = isFlagEnabled(ctx.featuresCohort, 'prismicArticlesOnExplore', flags);
+  const contentListPromise = prismicArticlesOnExploreFlag ? getEditorialList() : Promise.resolve([]);
+
+  const listRequests = [getCuratedList('explore'), getArticleStubs(10), contentListPromise];
+  const [curatedList, articleStubs, contentList] = await Promise.all(listRequests);
+
+  const wpPromos = articleStubs.data.map(PromoFactory.fromArticleStub);
+  const contentPromos = contentList.map(PromoFactory.fromArticleStub);
+  const promos = wpPromos.concat(contentPromos).sort((a, b) => {
+    return a.datePublished.getTime() - b.datePublished.getTime();
+  }).reverse();
+
+  // TODO: Remove this, make it automatic
+  const latestTweets = ctx.intervalCache.get('tweets');
+
+  ctx.render('pages/curated-lists', {
+    pageConfig: createPageConfig({
+      title: 'Explore',
+      inSection: 'explore',
+      category: 'list'
+    }),
+    promos,
+    curatedList,
+    collectorsPromo, // TODO: Remove this, make it automatic
+    latestTweets
+  });
+
+  return next();
+}
